+++ CHANGELOG.md
@@ -1,2 +1,4 @@
 # 1.11.7
 
+* Fixes for working on a machine with a low CPU count (better MaxDOP management for Python processes)
+* Current DQ status calculated across different table comparisons.